Defect #2723507 is fixed. Similar occurrences are found and fixed as well. Here is the description of the problem.
Never ever do the following, always set all date parts at once!

	date.setFullYear(year);
	date.setMonth(month-1);
	date.setDate(day);

This construct depends on execution time's day and can result in a wrong month being set.
i.e. say today is 31st of March 2009, we pass (2009, September, 11) and date will be set to 2009 October 11.
This happens because today's day is 31st and September only has 30 days
so date is normalized to 2009 October 01 when setMonth() is called.